+++ .buildkite/commands/prototype-upload.sh
@@ -10,9 +10,6 @@ buildkite-agent artifact download "artifacts/*.app.dSYM.zip" . --step build_prot
 echo "--- :rubygems: Setting up Gems"
 install_gems
 
-echo "--- :cocoapods: Setting up Pods"
-install_cocoapods
-
 echo "--- :closed_lock_with_key: Installing Secrets"
 bundle exec fastlane run configure_apply